Create README.md
My name is Anshul Teotia and I was born and Raised in Meerut (UP). 
I have completed my Bachelor of Computer Application With the agrecate of 62.5%. and after that I am persuing  my Master in Computer Application From Bennett University with the agrecate of 71.65.
I have strong foundation in python and focusing in Machine Learning and Computer Vision and Deep Learning.